rm opts for now
